requirements.txt
numpy>=1.24
opencv-python>=4.8
Pillow>=10.0
tensorflow>=2.15
torch>=2.2
torchvision>=0.17
ultralytics>=8.0
scikit-learn>=1.3
matplotlib>=3.7
tqdm>=4.66
tensorboard>=2.15
